Change In Note Placement System
Changed system from using the data to arbitrarily correlate to some number on the 360 degrees of the circle, to propagating notes along the circle evenly with a max amount of 25 notes per voice. This makes the sonic and visual experience more symmetrical.  I'll see if people enjoy that or not. The plus side is that it also makes it more obviously connected to the data.